Add keccak256 precompile (#17)
Part of #3.

Adds a way to call a built-in keccak through the vm.

I tested this by just calling the function through a few places in the
example contract, but I didn't include that code here because it doesn't
really make much sense.

I'm not sure how to test it properly otherwise, does it fit somewhere in
the existing contract or would we need a different test (but I don't see
any unit tests).

~Also I'm not sure if putting these in their own module makes sense, but
it just feels like the `env` module is too generic (or maybe it's just
the name that's confusing?).~ Changed my mind about this, although I
still think "env" is a weird name for the shared module.