Add permissioned token example, basic effects (#20)
Reference #18 

I tried to implement the basic logic with the existing system, but I
found a few things annoying or not possible, so I did put together a
pretty basic effect-handler system. There are some quirks like the fact
that I'm using strings as ids for the effects, which may not be ideal,
but it makes debugging a bit simpler. Also I did write an implementation
of `UtxoResume`, although I haven't tested it that much yet.

## Future work

The following are things I think shouldn't be in this PR mostly to limit
the scope a bit, since this is already quite big I think (although there
is a bit of boilerplate).

- The public keys in the blacklist/addresses for now are `i32` to
  simplify passing things around.
- Probably the effects should have some unique (content addressed?)
  identifier so that it's not just the name that matters?
- Reconciliate the effects with the current yield function. Both things
  are fairly similar, but the yield is currently a bit more limited. But
  the current yield has the advantage of the targeted queries (basically
  selecting the effect handler). With the implementation in this PR this
  would probably require handlers with id's combined with the utxo 
  handle or something like that.
- Add some way of distinguishing tokens by their type. Otherwise
  something like "get all the tokens of this utxo" is hard to handle.