Bootstrap ebnf grammar (#21)
Closes #19.

The syntax (of the grammar) is that of https://www.bottlecaps.de/rex/

Not to use that for the actual parser of course, just because it uses
w3c ebnf inspired syntax, and it was a quick way for me to test (and the
railroad graphs may be useful?)

Currently this is an LL(2) grammar. It's not really needed for the
design, and it makes some things more annoying, but I thought it may be
easier to translate it to the Rust parser this way.

I tried to keep the syntax in the slides, but there may be things I
missed. Obviously this is not final so feel free to suggest changes or
missing things.