Add chumsky parser (#23)
This adds a chumsky-based parser for the ebnf grammar as proposed in #21

The only difference should be that I used the pratt parser for the
operators. This is not required, but it's simpler to write and it gives
you correct associativity (with the ebnf grammar doesn't have since I
didn't write it with precedence climbing).

A few notes:

1. The parser currently gives you an AST and not a parse tree. It's
   possible that for some use-cases we may need to keep the tokens 
   around.
2. I didn't add any sort of error recovery or any sort of handling of
   error messages. I'm still not entirely sure how this works, so I'd 
   need to look into it a bit more.
3. Also didn't pay any attention to performance.