Add tree-sitter grammar and highlighting query (#28)
We don't necessarily need to merge this now since it may be annoying to
keep the grammar in sync while the language evolves (and in particular
until we have codegen writing in the dsl is not that useful anyway).

I mainly did set this up because I wanted syntax highlighting in helix
when working with the examples. It may be useful if someone wants to
re-use it for neovim/zed/emacs. Maybe it can be used in the sandbox too,
although that'll probably require more work.