Convert VSCode extension to a web extension (#65)
A VSCode [web extension] can run in VSCode's [browser mode] and can in
principle also be adapted to run in our own browser sandbox.

To achieve this, the new `starstream-language-server-web` crate uses
`web-sys` and `wasm-bindgen-futures` to adapt the language server to run
in a Web Worker. It receives `postMessage({ "jsonrpc": 2.0, ... })`
calls and posts replies back directly, without going through a byte
stream or even JSON.

The language server itself changed a bit:
- Removed `initialized: AtomicBool` since tower-lsp-server already
  rejects double-initialization.
- Commented out `workspace_folders` for now since
  `tokio::sync::OnceCell` is not available on the web. An alternative
  interior mutability may be needed.
- Wrapped `LspService::new` to avoid having to update two places if we
  ever add custom methods.

Quirks on the VSCode side:
- Webpack's URL handling doesn't work because scripts are loaded under a
  `blob:`-like URL.
- So we use `wasm-bindgen --mode web` to get an `init` function that we
  can manually pass the Wasm blob to.
- Build depends on `cargo`, `wasm-bindgen`, and `wasm-opt` without
  declaration.

I'm slightly worried about the possibility of a race in the web worker
initialization but haven't confirmed/denied it yet.

[web extension]: https://code.visualstudio.com/api/extension-guides/web-extensions
[browser mode]: https://vscode.dev/

---------

Signed-off-by: Tad Hardesty <tad.hardesty@midnight.foundation>