Snapshot test all outputs of main test cases (#93)
For each program in the input corpus, snapshots:
- AST
- Inference trace
- Typed AST
- Core WebAssembly
- WIT

Includes warnings and stop if any stage returns a fatal error. Also
asserts that the input file is correctly formatted. This way the whole
compiler is exercised for each input file.

Also moves the whole-program parse cases out of `starstream-compiler`
`#[test]`s and into the above system.

---------

Signed-off-by: Tad Hardesty <tad.hardesty@midnight.foundation>