Implement component ABI for functions returning records (#100)
Follows #99 by implementing the component ABI's scheme of writing to
linear memory when a function would otherwise be returning multiple
results.

The `component_abi` module is expanded to have a `ComponentAbiType` enum
from which size, alignment, and store operations can be computed.

The Wasm compiler will now export a memory for the component bindings to
read from.